archive/ai-vault-planning-cli-1/pyproject.toml
[tool]
name = "ai-vault-planning-cli"
version = "0.1.0"
description = "A CLI for managing AI-assisted vault planning."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/ai-vault-planning-cli"
repository = "https://github.com/yourusername/ai-vault-planning-cli"
keywords = ["AI", "CLI", "vault", "planning", "markdown"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = "^3.8"
sentence-transformers = "^2.2.0"
openai = "^0.11.0"
pyyaml = "^5.4.1"
faiss-cpu = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.7b0"
flake8 = "^3.9.2"
mypy = "^0.910"